64-bit porting
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1016 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/include/G3D/Array.h
@@ -856,7 +856,7 @@ public:
     inline T& randomElement() {
         debugAssert(num > 0);
         debugAssert(data!=NULL);
-        return data[iRandom(0, num - 1)];
+        return data[iRandom(0, (int)num - 1)];
     }
 
     inline const T& randomElement() const {
